+++ examples/onnx/Cargo.toml
@@ -9,6 +9,7 @@ guest = { package = "onnx-guest", path = "./guest" }
 tract-onnx = "0.21.1"
 ndarray = "0.16.1"
 anyhow = "1.0.75"
+onnx-util = { package = "onnx-util", path = "./util" }
 
 [features]
 icicle = ["jolt-sdk/icicle"]
