+++ examples/onnx/guest/Cargo.toml
@@ -8,8 +8,4 @@ guest = []
 
 [dependencies]
 jolt = { package = "jolt-sdk", path = "../../../jolt-sdk", features = ["guest-std"] }
-tract-onnx = "0.21.1"
-anyhow = "1.0.75"
-ndarray = "0.16.1"
-getrandom = { version = "0.2", features = ["custom"] }
-
+onnx-util = { package = "onnx-util", path = "../util", default-features = false, features = ["guest"] }
